+++ VerstaOrders/Model/Dto/OrderDto.cs
@@ -6,5 +6,5 @@ public record OrderDto(
     string AddressSender,
     string TownReceiver,
     string AddressReceiver,
-    double ProductWeight,
+    decimal ProductWeight,
     DateTime PickupDate);
